refactor: Move `log_clock` to core subfolder
