+++ libs/deeplog/src/dplx/dlog/core/log_clock.test.cpp
@@ -1,0 +1,8 @@
+
+// Copyright Henrik Steffen Gaßmann 2021
+//
+// Distributed under the Boost Software License, Version 1.0.
+//         (See accompanying file LICENSE or copy at
+//           https://www.boost.org/LICENSE_1_0.txt)
+
+#include "dplx/dlog/core/log_clock.hpp"
