+++ libs/deeplog_tui/src/dplx/dlog/tui/log_display_grid.hpp
@@ -14,8 +14,8 @@
 #include <ftxui/component/component_base.hpp>
 #include <ftxui/component/event.hpp>
 
+#include <dplx/dlog/core/log_clock.hpp>
 #include <dplx/dlog/definitions.hpp>
-#include <dplx/dlog/log_clock.hpp>
 #include <dplx/dlog/record_container.hpp>
 #include <dplx/dlog/tui/theme.hpp>
